Added signup.html and account.html
Added a sign up button to the login page that redirects to the account creation page. Changed the style of login.html, saveduniversities.html, and search-page-results-list-2.0.html to match the index.html. Created a dummy account to login to. Added a profile icon to the top right of index.html