Update deploy.yml
https://github.com/marketplace/actions/generate-snake-game-from-github-contribution-grid

This action was found via the above link.

From the website:

"Pull a github user's contribution graph. Make it a snake Game, generate a snake path where the cells get eaten in an orderly fashion... Available as github action. It can automatically generate a new image each day. Which makes for great github profile readme"

---------------------------------------------------------------------

Note:

I was uneasy about committing directly to the "hsowda" branch, which was created for the sake of editing this "deploy.yml" file. Because of this, I have decided to create a new branch called: "hsowda-patch-14" that I will use to propose my changes on, just to be on the safe side.

---------------------------------------------------------------------

11.27.2024  |  12:45AM